requirements.txt
# Core
streamlit>=1.20
pandas>=1.5
plotly>=5.15

# NLP & parsing
spacy>=3.5.4
sentence-transformers>=2.2.2
pdfminer.six>=20201018
python-docx>=0.8.11
chardet>=5.0.0

# DB & utils
sqlalchemy>=1.4
pydantic
numpy
scikit-learn
tqdm

# Use the modern Google GenAI SDK
google-genai>=0.9.0
